% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\docType{data}
\name{Goolam}
\alias{Goolam}
\title{Goolam}
\format{
An object of class \code{list} of length 2.
}
\usage{
Goolam
}
\description{
Goolam dataset with data and cell types information.The number of genes is reduced to 10,000.
}
\keyword{datasets}
