% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscISR.R
\name{plotscISR}
\alias{plotscISR}
\title{plotscISR: Visualization of scRNA-seq data.}
\usage{
plotscISR(data, label, perplexity = 30)
}
\arguments{
\item{data}{Input matrix or data frame. Rows represent genes while columns represent samples.}

\item{label}{A vector that contains the cell type labels.}

\item{perplexity}{Number of closed neighbors for each data point.}
}
\value{
\code{plotscISR} Return the visualization of the cell populations in the scRNA-seq dataset.
}
\description{
Plotting scRNA-seq data to visualize the cell transciptome landscape.
}
\details{
This function utilizes the dimension reduction feature of t-SNE package to viusualize high throughput data over two dimensional space.
}
\examples{

# Load the sample dataset scISRExample
data(scISRExample)

# Perform the imputation
imputed <- scISR(data = scISRExample$dropout)

# Plot the complete data
plot_raw <- plotscISR(scISRExample$raw, label = scISRExample$celltype)

# Plot the dropout data
plot_dropout <- plotscISR(scISRExample$dropout, label = scISRExample$celltype)

# Plot the imputed data
plot_imputed <- plotscISR(imputed, label = scISRExample$celltype)

}
