% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scISR.R
\name{scISR}
\alias{scISR}
\title{scISR: Single-cell Imputation using Subspace Regression}
\usage{
scISR(data, ncores = 1, seed = 1)
}
\arguments{
\item{data}{Input matrix or data frame. Rows represent genes while columns represent samples}

\item{ncores}{Seed for reproducibility. Default value is \code{1}.}
}
\value{
\code{scISR} returns an imputed single-cell expression matrix where rows represent genes while columns represent samples.
}
\description{
Perform single-cell Imputation using Subspace Regression
}
\details{
scISR performs imputation for single-cell sequencing data. scISR identifies the true dropout values in the scRNA-seq dataset using
hyper-geomtric testing approach. Based on the result obtained from hyper-geometric testing, the original dataset is segregated into two
subsets including training data and imputable data. Next, training data is used for constructing a generalize linear regression model that
is used for imputation on the imputable data.
}
\examples{

# Load the sample dataset scISRExample
data(scISRExample)

# Perform the imputation
imputed <- scISR(data = scISRExample$dropout)

# Plot the complete data
plotscISR(scISRExample$raw, label = scISRExample$celltype)

# Plot the dropout data
plotscISR(scISRExample$dropout, label = scISRExample$celltype)

# Plot the imputed data
plotscISR(imputed, label = scISRExample$celltype)

}
\references{
1. Nguyen, T., Tagett, R., Diaz, D., & Draghici, S. (2017). A novel approach for data integration and disease subtyping. Genome research, 27(12), 2025-2039.
}
